+++ backend/src/auth/auth.module.ts
@@ -3,11 +3,23 @@ import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
 import { MailModule } from '../mail/mail.module';
 import { UserModule } from '../user/user.module';
+import { JwtModule } from '@nestjs/jwt';
+import { FileModule } from '../file/file.module';
 
 @Module({
-    imports: [MailModule, forwardRef(() => UserModule)],
+    imports: [
+        MailModule,
+        FileModule,
+        forwardRef(() => UserModule),
+        JwtModule.register({
+            secret: process.env.PRIVATE_KEY || 'SECRET',
+            signOptions: {
+                expiresIn: '1h',
+            },
+        }),
+    ],
     controllers: [AuthController],
     providers: [AuthService],
-    exports: [AuthService],
+    exports: [AuthService, JwtModule],
 })
 export class AuthModule {}