+++ backend/src/user/user.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Post, Body, Param, Get } from '@nestjs/common';
+import { Controller, Post, Body, Param, Get, Delete, UploadedFile } from '@nestjs/common';
 import { UserService } from './user.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { AddRoleDto } from './dto/add-role.dto';
@@ -13,12 +13,12 @@ export class UserController {
     }
 
     @Post()
-    create(@Body() createUserDto: CreateUserDto) {
+    create( @UploadedFile()@Body() createUserDto: CreateUserDto) {
         return this.usersService.create(createUserDto);
     }
 
-    @Post('/role')
+    @Post('/role:id')
     addRole(@Param('id') id: number, @Body() dto: AddRoleDto) {
-        return this.usersService.addRole(id ,dto)
+        return this.usersService.addRole(id, dto);
     }
 }