+++ backend/src/user/user.service.ts
@@ -1,28 +1,32 @@
 import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
-import { UpdateUserDto } from './dto/update-user.dto';
 import { InjectModel } from '@nestjs/sequelize';
-import { UserModule } from './user.module';
 import { UserModel } from './models/user.model';
 import * as uuid from 'uuid';
 import { MailService } from '../mail/mail.service';
 import { RoleService } from '../role/role.service';
 import { AddRoleDto } from './dto/add-role.dto';
+import { FileService } from '../file/file.service';
 
 @Injectable()
 export class UserService {
     constructor(
-        @InjectModel(UserModel) private userRepository: typeof UserModel,
-        private mailService: MailService, private roleService: RoleService,
+        @InjectModel(UserModel)
+        private userRepository: typeof UserModel,
+        private mailService: MailService,
+        private roleService: RoleService,
+        private fileService: FileService,
     ) {
     }
 
     async create(userDto: CreateUserDto) {
+        const fileName = await this.fileService.createFile(userDto.avatar);
         const activationLink = uuid.v4();
         const user = await this.userRepository.create(userDto);
         const role = await this.roleService.getRoleByValue('USER');
         await this.mailService.sendActivationMail(userDto.email, `http://localhost:5000/auth/activate/${activationLink}`);
         user.activationLink = activationLink;
+        user.avatar = fileName;
         await user.$set('roles', [role.id]);
         user.roles = [role];
         await user.save();
@@ -30,7 +34,7 @@ export class UserService {
     }
 
     async getUsers() {
-       return await this.userRepository.findAll({ include: { all: true }})
+        return await this.userRepository.findAll({ include: { all: true } });
     }
 
     async addRole(id: number, dto: AddRoleDto) {
@@ -44,6 +48,9 @@ export class UserService {
     }
 
     async getByEmail(email: string) {
-        return await this.userRepository.findOne({ where: { email }, include: { all: true }, });
+        return await this.userRepository.findOne({
+            where: { email },
+            include: { all: true },
+        });
     }
 }